# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.CreateTodoDTO:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.Todo:
    properties:
      completed:
        type: boolean
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      userId:
        description: Link to the User
        type: string
    required:
    - title
    type: object
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.UpdateTodoDTO:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        type: string
    type: object
...
info:
  contact:
    email: innocent@altschoolafrica.com
    name: API Support - Innocent
    url: https://github.com/Innocent9712
  description: This is an API for MuchToDo application with user authentication.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MuchToDo API
  version: "1.0"
paths:
...
  /todos:
    get:
      description: Retrieves a list of all todo items belonging to the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.Todo'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all todos for the current user
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Adds a new todo item to the current user's list
      parameters:
      - description: Todo Create Object
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.CreateTodoDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.Todo'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new todo
      tags:
      - todos
  /todos/{id}:
    delete:
      description: Deletes a specific todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{''message'': ''Todo deleted successfully''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Todo not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a todo
      tags:
      - todos
    get:
      description: Retrieves a specific todo item by its ID, if it belongs to the
        user
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.Todo'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Todo not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get a single todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Updates the details of a specific todo item
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Todo Update Object
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.UpdateTodoDTO'
      produces:
      - application/json
      responses:
        "200":
          description: '{''message'': ''Todo updated successfully''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Todo not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a todo
      tags:
      - todos
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer'' followed by a space and a JWT token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"